% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubist.R
\name{cubist}
\alias{cubist}
\alias{update.cubist}
\title{General Interface for Cubist Rule-Based Regression Models}
\usage{
cubist(mode = "regression", committees = NULL, neighbors = NULL)

\method{update}{cubist}(
  object,
  parameters = NULL,
  committees = NULL,
  neighbors = NULL,
  fresh = FALSE,
  ...
)
}
\arguments{
\item{mode}{A single character string for the type of model.
The only possible value for this model is "regression".}

\item{committees}{A non-negative integer (no greater than 100 for the number
of members of the ensemble.}

\item{neighbors}{An integer between zero and nine for the number of training
set instances that are used to adjust the model-based prediction.}

\item{object}{A Cubist model specification.}
}
\description{
\code{\link[=cubist]{cubist()}} is a way to generate a \emph{specification} of a model
before fitting. The main arguments for the model are:
\itemize{
\item \code{committees}: The number of sequential models included in the
ensemble (similar to the number of trees in boosting).
\item \code{neighbors}: The number of neighbors in the post-model
instance-based adjustment.
}
These arguments are converted to their specific names at the
time that the model is fit. Other options and argument can be
set using \code{\link[parsnip:set_engine]{parsnip::set_engine()}}. If left to their defaults
here (\code{NULL}), the values are taken from the underlying model
functions. If parameters need to be modified, \code{update()} can be used
in lieu of recreating the object from scratch.
}
\details{
The only availible engine is \code{"Cubist"}.
}
\section{Engine Details}{


Engines may have pre-set default arguments when executing the
model fit call. For this type of
model, the template of the fit calls are:

\pkg{Cubist}
}

\examples{
cubist()
# Parameters can be represented by a placeholder:
cubist(committees = 7)
model <- cubist(committees = 10, neighbors = 2)
model
update(model, committees = 1)
update(model, committees = 1, fresh = TRUE)
}
\seealso{
\code{\link[parsnip:fit]{parsnip::fit()}}, \code{\link[parsnip:set_engine]{parsnip::set_engine()}}
}
