% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_fit.R
\name{mtry_prop}
\alias{mtry_prop}
\title{Proportion of Randomly Selected Predictors}
\usage{
mtry_prop(range = c(0.1, 1), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::log10_trans()} or \code{scales::reciprocal_trans()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\value{
A \code{dials} with classes "quant_param" and "param". The \code{range} element
of the object is always converted to a list with elements "lower" and "upper".
}
\description{
Proportion of Randomly Selected Predictors
}
