% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{multi_predict}
\alias{tidy}
\alias{tunable}
\alias{mtry_prop}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dials}{\code{\link[dials:mtry_prop]{mtry_prop()}}}

  \item{generics}{\code{\link[generics:tidy]{tidy()}}, \code{\link[generics:tunable]{tunable()}}}

  \item{parsnip}{\code{\link[parsnip:multi_predict]{multi_predict()}}}
}}

