% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubist_rules.R, R/rule_fit.R
\name{multi_predict._cubist}
\alias{multi_predict._cubist}
\alias{multi_predict._xrf}
\title{\code{multi_predict()} methods for rule-based models}
\usage{
\method{multi_predict}{`_cubist`}(object, new_data, type = NULL, neighbors = NULL, ...)

\method{multi_predict}{`_xrf`}(object, new_data, type = NULL, penalty = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{model_fit}}

\item{new_data}{A rectangular data object, such as a data frame.}

\item{type}{A single character value or \code{NULL}.}

\item{neighbors}{An numeric vector of neighbors values between zero and nine.}

\item{...}{Not currently used.}

\item{penalty}{Non-negative penalty values.}
}
\description{
\code{multi_predict()} methods for rule-based models
}
