% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C5_rules.R, R/cubist.R, R/rule_fit.R
\name{multi_predict._C5.0}
\alias{multi_predict._C5.0}
\alias{multi_predict._cubist}
\alias{multi_predict._xrf}
\title{\code{\link[=multi_predict]{multi_predict()}} methods for rule-based models}
\usage{
\method{multi_predict}{`_C5.0`}(object, new_data, type = NULL, trees = NULL, ...)

\method{multi_predict}{`_cubist`}(object, new_data, type = NULL, neighbors = NULL, ...)

\method{multi_predict}{`_xrf`}(object, new_data, type = NULL, penalty = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{model_fit}}

\item{new_data}{A rectangular data object, such as a data frame.}

\item{type}{A single character value or \code{NULL}. Possible values
are class" and "prob".}

\item{trees}{An numeric vector of \code{trees} between one and 100.}

\item{...}{Not currently used.}

\item{neighbors}{An numeric vector of neighbors values between zero and nine.}

\item{penalty}{Non-negative penalty values.}
}
\description{
\code{\link[=multi_predict]{multi_predict()}} methods for rule-based models
}
