% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C5_rules.R, R/cubist.R
\name{multi_predict._C5.0}
\alias{multi_predict._C5.0}
\alias{multi_predict._cubist}
\title{\code{\link[=multi_predict]{multi_predict()}} methods for rule-based models}
\usage{
\method{multi_predict}{`_C5.0`}(object, new_data, type = NULL, trees = NULL, ...)

\method{multi_predict}{`_cubist`}(object, new_data, type = NULL, neighbors = NULL, ...)
}
\arguments{
\item{trees}{An numeric vector of \code{trees} between one and 100.}

\item{neighbors}{An numeric vector of neighbors values between zero and nine.}
}
\description{
\code{\link[=multi_predict]{multi_predict()}} methods for rule-based models
}
