% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C5_rules.R, R/cubist_rules.R, R/rule_fit.R
\name{c5_fit}
\alias{c5_fit}
\alias{c5_pred}
\alias{tunable.C5_rules}
\alias{cubist_fit}
\alias{get_neighbors}
\alias{tunable.cubist_rules}
\alias{xrf_fit}
\alias{xrf_pred}
\alias{qrf_fit}
\alias{qrf_pred}
\alias{multi_predict._qrf}
\title{Internal function wrappers}
\usage{
c5_fit(x, y, trials = 1, minCases = 2, cost = NULL, ...)

c5_pred(object, new_data, trials = object$fit$trials["Actual"], ...)

\method{tunable}{C5_rules}(x, ...)

cubist_fit(x, y, committees = 1, neighbors = 0, max_rules = NA, ...)

get_neighbors(x)

\method{tunable}{cubist_rules}(x, ...)

xrf_fit(
  formula,
  data,
  max_depth = 6,
  nrounds = 15,
  eta = 0.3,
  colsample_bynode = NULL,
  colsample_bytree = NULL,
  min_child_weight = 1,
  gamma = 0,
  subsample = 1,
  validation = 0,
  early_stop = NULL,
  counts = TRUE,
  event_level = c("first", "second"),
  lambda = 0.1,
  objective = NULL,
  ...
)

xrf_pred(object, new_data, lambda = object$fit$lambda, type, ...)

qrf_fit(x, y, weights = NULL, iterations = 5, penalty = 0.01, mixture = 1, ...)

qrf_pred(object, new_data, type, ...)

\method{multi_predict}{`_qrf`}(object, new_data, type = NULL, penalty = NULL, ...)
}
\description{
These are not supported when called by the user.
}
\keyword{internal}
