% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C5_rules.R, R/cubist.R, R/rule_fit.R
\name{c5_fit}
\alias{c5_fit}
\alias{c5_pred}
\alias{cubist_fit}
\alias{tunable.cubist_rules}
\alias{xrf_fit}
\alias{xrf_pred}
\alias{tunable.rule_fit}
\title{Internal function wrappers}
\usage{
c5_fit(x, y, trials = 1, minCases = 2, cost = NULL, ...)

c5_pred(object, new_data, trials = object$fit$trials["Actual"], ...)

cubist_fit(x, y, committees = 1, neighbors = 0, max_rules = NA, ...)

tunable.cubist_rules(x, ...)

xrf_fit(
  formula,
  data,
  max_depth = 6,
  nrounds = 15,
  eta = 0.3,
  colsample_bytree = 1,
  min_child_weight = 1,
  gamma = 0,
  subsample = 1,
  lambda = 0.1,
  ...
)

xrf_pred(object, new_data, lambda = object$fit$lambda, type, ...)

tunable.rule_fit(x, ...)
}
\description{
These are not supported when called by the user.
}
\keyword{internal}
