% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules-package.R
\docType{package}
\name{rules-package}
\alias{rules}
\alias{rules-package}
\title{rules: Model Wrappers for Rule-Based Models}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Bindings for additional models for use with the 'parsnip' package. Models include prediction rule ensembles (Friedman and Popescu, 2008) \doi{10.1214/07-AOAS148}, C5.0 rules (Quinlan, 1992 ISBN: 1558602380), and Cubist (Kuhn and Johnson, 2013) \doi{10.1007/978-1-4614-6849-3}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidymodels/rules}
  \item \url{https://rules.tidymodels.org/}
  \item Report bugs at \url{https://github.com/tidymodels/rules/issues}
}

}
\author{
\strong{Maintainer}: Emil Hvitfeldt \email{emilhhvitfeldt@gmail.com} (\href{https://orcid.org/0000-0002-0679-1945}{ORCID})

Authors:
\itemize{
  \item Max Kuhn \email{max@rstudio.com} (\href{https://orcid.org/0000-0003-2402-136X}{ORCID})
}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
