% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_C5.0.R, R/tidy_cubist.R, R/tidy_xrf.R
\name{tidy.C5.0}
\alias{tidy.C5.0}
\alias{tidy.cubist}
\alias{tidy.xrf}
\title{Turn rule models into tidy tibbles}
\usage{
\method{tidy}{C5.0}(x, trials = x$trials["Actual"], ...)

\method{tidy}{cubist}(x, committees = x$committee, ...)

\method{tidy}{xrf}(x, penalty = NULL, unit = c("rules", "columns"), ...)
}
\arguments{
\item{x}{A \code{Cubist}, \code{C5.0}, or \code{xrf} object.}

\item{trials}{The number of boosting iterations to tidy (defaults to the entire
ensemble).}

\item{...}{Not currently used.}

\item{committees}{The number of committees to tidy (defaults to the entire
ensemble).}

\item{penalty}{A single numeric value for the \code{lambda} penalty value.}

\item{unit}{What data should be returned? For \code{unit = 'rules'}, each row
corresponds to a rule. For \code{unit = 'columns'}, each row is a predictor
column. The latter can be helpful when determining variable importance.}
}
\value{
The Cubist method has columns \code{committee}, \code{rule_num}, \code{rule}, \code{estimate},
and \code{statistic}. The latter two are nested tibbles. \code{estimate} contains
the parameter estimates for each term in the regression model and \code{statistic}
has statistics about the data selected by the rules and the model fit.

The C5.0 method has columns \code{trial}, \code{rule_num}, \code{rule},
and \code{statistics}. The latter two are nested tibbles.  \code{statistic}
has statistics about the data selected by the rules.

The \code{xrf} results has columns \code{rule_id}, \code{rule}, and \code{estimate}. The
\code{rule_id} column has the rule identifier (e.g., "r0_21") or the feature
column name when the column is added directly into the model. For multiclass
models, a \code{class} column is included.

In each case, the \code{rule} column has a character string with the rule
conditions. These can be converted to an R expression using
\code{\link[rlang:parse_expr]{rlang::parse_expr()}}.
}
\description{
Turn rule models into tidy tibbles
}
\details{
\subsection{An example}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(dplyr)
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{## 
## Attaching package: 'dplyr'

## The following objects are masked from 'package:stats':
## 
##     filter, lag

## The following objects are masked from 'package:base':
## 
##     intersect, setdiff, setequal, union
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{data(ames, package = "modeldata")

ames <-
  ames \%>\%
  mutate(Sale_Price = log10(ames$Sale_Price),
         Gr_Liv_Area = log10(ames$Gr_Liv_Area))

# ------------------------------------------------------------------------------

cb_fit <-
  cubist_rules(committees = 10) \%>\%
  set_engine("Cubist") \%>\%
  fit(Sale_Price ~ Neighborhood + Longitude + Latitude + Gr_Liv_Area + Central_Air,
      data = ames)

cb_res <- tidy(cb_fit)
cb_res
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{## # A tibble: 157 x 5
##    committee rule_num rule                                     estimate statis~1
##        <int>    <int> <chr>                                    <list>   <list>  
##  1         1        1 ( Central_Air == 'N' ) & ( Gr_Liv_Area ~ <tibble> <tibble>
##  2         1        2 ( Gr_Liv_Area <= 3.0326188 ) & ( Neighb~ <tibble> <tibble>
##  3         1        3 ( Neighborhood  \%in\% c( 'Old_Town','Edw~ <tibble> <tibble>
##  4         1        4 ( Neighborhood  \%in\% c( 'Old_Town','Edw~ <tibble> <tibble>
##  5         1        5 ( Central_Air == 'N' ) & ( Gr_Liv_Area ~ <tibble> <tibble>
##  6         1        6 ( Longitude <= -93.652023 ) & ( Neighbo~ <tibble> <tibble>
##  7         1        7 ( Gr_Liv_Area > 3.2284005 ) & ( Neighbo~ <tibble> <tibble>
##  8         1        8 ( Neighborhood  \%in\% c( 'North_Ames','G~ <tibble> <tibble>
##  9         1        9 ( Latitude <= 42.009399 ) & ( Neighborh~ <tibble> <tibble>
## 10         1       10 ( Neighborhood  \%in\% c( 'College_Creek'~ <tibble> <tibble>
## # ... with 147 more rows, and abbreviated variable name 1: statistic
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{cb_res$estimate[[1]]
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{## # A tibble: 4 x 2
##   term        estimate
##   <chr>          <dbl>
## 1 (Intercept)  -408.  
## 2 Longitude      -1.43
## 3 Latitude        6.6 
## 4 Gr_Liv_Area     0.7
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{cb_res$statistic[[1]]
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{## # A tibble: 1 x 6
##   num_conditions coverage  mean   min   max  error
##            <dbl>    <dbl> <dbl> <dbl> <dbl>  <dbl>
## 1              2      154  4.94  4.11  5.31 0.0956
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# ------------------------------------------------------------------------------

library(recipes)
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{## 
## Attaching package: 'recipes'

## The following object is masked from 'package:stats':
## 
##     step
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{xrf_reg_mod <-
  rule_fit(trees = 10, penalty = .001) \%>\%
  set_engine("xrf") \%>\%
  set_mode("regression")

# Make dummy variables since xgboost will not
ames_rec <-
  recipe(Sale_Price ~ Neighborhood + Longitude + Latitude +
         Gr_Liv_Area + Central_Air,
         data = ames) \%>\%
  step_dummy(Neighborhood, Central_Air) \%>\%
  step_zv(all_predictors())

ames_processed <- prep(ames_rec) \%>\% bake(new_data = NULL)

set.seed(1)
xrf_reg_fit <-
  xrf_reg_mod \%>\%
  fit(Sale_Price ~ ., data = ames_processed)

xrf_rule_res <- tidy(xrf_reg_fit, penalty = 0.05)
xrf_rule_res$rule[nrow(xrf_rule_res)] \%>\% rlang::parse_expr()
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{## (Latitude < 42.0367355) & (Latitude >= 42.0094452) & (Neighborhood_College_Creek < 
##     0.5) & (Neighborhood_Northridge_Heights < 0.5) & (Neighborhood_Somerset < 
##     0.5) & (Neighborhood_Stone_Brook < 0.5)
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{xrf_col_res <- tidy(xrf_reg_fit, penalty = 0.05, unit = "columns")
xrf_col_res
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{## # A tibble: 19 x 3
##    rule_id       term                                                  estimate
##    <chr>         <chr>                                                    <dbl>
##  1 r3_5          Longitude                                             0.00731 
##  2 r3_2          Longitude                                            -0.0101  
##  3 r3_5          Latitude                                              0.00731 
##  4 r3_5          Longitude                                             0.00731 
##  5 r7_47         Gr_Liv_Area                                           0.00186 
##  6 r7_47         Central_Air_Y                                         0.00186 
##  7 r7_47         Longitude                                             0.00186 
##  8 r7_47         Neighborhood_Edwards                                  0.00186 
##  9 r7_47         Neighborhood_Sawyer                                   0.00186 
## 10 r7_47         Neighborhood_South_and_West_of_Iowa_State_University  0.00186 
## 11 r8_25         Neighborhood_Northridge_Heights                      -0.0290  
## 12 r8_25         Neighborhood_Stone_Brook                             -0.0290  
## 13 r8_25         Latitude                                             -0.0290  
## 14 r8_25         Neighborhood_Somerset                                -0.0290  
## 15 r8_25         Neighborhood_College_Creek                           -0.0290  
## 16 r8_25         Latitude                                             -0.0290  
## 17 (Intercept)   (Intercept)                                           3.59    
## 18 Gr_Liv_Area   Gr_Liv_Area                                           0.521   
## 19 Central_Air_Y Central_Air_Y                                         0.000314
}\if{html}{\out{</div>}}
}
}
