% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_C5.0.R, R/tidy_cubist.R, R/tidy_xrf.R
\name{tidy.C5.0}
\alias{tidy.C5.0}
\alias{tidy.cubist}
\alias{tidy.xrf}
\title{Turn C5.0 and rule-based models into tidy tibbles}
\usage{
\method{tidy}{C5.0}(x, trees = x$trials["Actual"], ...)

\method{tidy}{cubist}(x, committees = x$committee, ...)

\method{tidy}{xrf}(x, penalty = NULL, unit = c("rules", "columns"), ...)
}
\arguments{
\item{x}{A \code{Cubist}, \code{C5.0}, or \code{xrf} object.}

\item{trees}{The number of boosting iterations to tidy (defaults to the entire
ensemble).}

\item{...}{Not currently used.}

\item{committees}{The number of committees to tidy (defaults to the entire
ensemble).}

\item{penalty}{A single numeric value for the \code{lambda} penalty value.}

\item{unit}{What data should be returned? For \code{unit = 'rules'}, each row
corresponds to a rule. For \code{unit = 'columns'}, each row is a predictor
column. The latter can be helpful when determining variable importance.}
}
\value{
\subsection{Cubist}{

The Cubist method has columns \code{committee}, \code{rule_num}, \code{rule}, \code{estimate},
and \code{statistic}. The latter two are nested tibbles. \code{estimate} contains
the parameter estimates for each term in the regression model and \code{statistic}
has statistics about the data selected by the rules and the model fit.
}

\subsection{C5.0}{

C5.0 models have different results depending on the type of model produced
by \code{\link[C50:C5.0]{C50::C5.0()}}.

For rule-based models, the resulting tibble has columns \code{tree}, \code{rule_num},
\code{rule}, and \code{statistics}. The latter two are nested tibbles.  \code{statistic}
has statistics about the data selected by the rules.

Tree models, including booted trees, are not stored in a nested structure but
in rules. For this model, the rule defines the variables and values that
define the splits all the way down to the terminal node. The columns in the
tibble are \code{tree}, \code{node}, \code{rule}, and \code{statistics} (a nested tibble).
}

\subsection{RuleFit}{

The \code{xrf} results has columns \code{rule_id}, \code{rule}, and \code{estimate}. The
\code{rule_id} column has the rule identifier (e.g., "r0_21") or the feature
column name when the column is added directly into the model. For multiclass
models, a \code{class} column is included.

In each case, the \code{rule} column has a character string with the rule
conditions. These can be converted to an R expression using
\code{\link[rlang:parse_expr]{rlang::parse_expr()}}.
}
}
\description{
Turn C5.0 and rule-based models into tidy tibbles
}
